feat: add transaction handler

+++ src/users/user.handler.go
@@ -1,9 +1,11 @@
 package users
 
+import "C"
 import (
+	"fiber/src/common/errors"
 	"github.com/gofiber/fiber/v2"
 	"github.com/google/wire"
-	"github.com/sirupsen/logrus"
+	"gorm.io/gorm"
 	"strconv"
 )
 
@@ -26,11 +28,12 @@ func NewUserHandler(service UserService) UserHandler {
 var SetHandler = wire.NewSet(NewUserHandler)
 
 func (handler userHandler) CreateOne(c *fiber.Ctx) error {
+	tx := c.Locals("TX").(*gorm.DB)
 	user := new(User)
 	if err := c.BodyParser(user); err != nil {
 		return c.Status(fiber.StatusBadRequest).SendString(err.Error())
 	}
-	result, err := handler.service.CreateOne(user)
+	result, err := handler.service.WithTx(tx).CreateOne(user)
 	if err != nil {
 		return err
 	}
@@ -46,17 +49,24 @@ func (handler userHandler) FindMany(c *fiber.Ctx) error {
 }
 
 func (handler userHandler) FindOne(c *fiber.Ctx) error {
-	id, _ := strconv.Atoi(c.Params("id"))
+	id, parseErr := strconv.Atoi(c.Params("id"))
+	if parseErr != nil {
+		return errors.HandleParseError(c, parseErr)
+	}
+
 	result, err := handler.service.FindOne(id)
 	if err != nil {
 		return err
 	}
-	logrus.Info("test")
 	return c.Status(fiber.StatusOK).JSON(result)
 }
 
 func (handler userHandler) UpdateOne(c *fiber.Ctx) error {
-	id, _ := strconv.Atoi(c.Params("id"))
+	id, parseErr := strconv.Atoi(c.Params("id"))
+	if parseErr != nil {
+		return errors.HandleParseError(c, parseErr)
+	}
+
 	user := new(User)
 	if err := c.BodyParser(user); err != nil {
 		return c.Status(fiber.StatusBadRequest).SendString(err.Error())
@@ -70,7 +80,11 @@ func (handler userHandler) UpdateOne(c *fiber.Ctx) error {
 }
 
 func (handler userHandler) DeleteOne(c *fiber.Ctx) error {
-	id, _ := strconv.Atoi(c.Params("id"))
+	id, parseErr := strconv.Atoi(c.Params("id"))
+	if parseErr != nil {
+		return errors.HandleParseError(c, parseErr)
+	}
+
 	user, err := handler.service.DeleteOne(id)
 	if err != nil {
 		return err
